fix bug

+++ app/src/main/java/com/breadwallet/wallet/wallets/ela/request/Outputs.java
@@ -4,6 +4,6 @@ public class Outputs {
 
     public String addr;
 
-    public int amt;
+    public long amt;
 
 }
